refactor(frontend): extract user and permission api clients
