+++ frontend/src/lib/services.ts
@@ -5,7 +5,6 @@ import {
    DOCUMENTS_ENDPOINT,
    HEALTH_ENDPOINT,
    LANGUAGE_ENDPOINT,
-   PERMISSIONS_ENDPOINT,
    PERSPECTIVES_ENDPOINT,
    TAGS_ENDPOINT,
    USERS_ENDPOINT
@@ -294,45 +293,26 @@ export function useTagsStatus(cid: string) {
    return { status: data, error, refresh: mutate };
 }
 
-export const registerUser = async (data: Record<string, string>) => {
-   const url = `${USERS_ENDPOINT}/register`;
-   const options: RequestInit = {
-      method: "POST",
-      headers: {
-         "Content-Type": "application/json"
-      },
-      body: JSON.stringify({
-         name: data.name,
-         email: data.email,
-         password: data.password,
-         confirmPassword: data.confirmPassword,
-         lang: data.lang,
-         confirmationLink: data.confirmationLink
-      }),
-      credentials: "include"
-   };
-   const response = await fetchWithCredentials(url, withCsrf(options));
-   const result = await response.json();
-   return result;
-};
-
-export const loginUser = async (data: Record<string, string>) => {
-   const url = `${USERS_ENDPOINT}/login`;
-   const options: RequestInit = {
-      method: "POST",
-      headers: {
-         "Content-Type": "application/json"
-      },
-      body: JSON.stringify({
-         email: data.email,
-         password: data.password
-      }),
-      credentials: "include"
-   };
-   const response = await fetchWithCredentials(url, withCsrf(options));
-   const result = await response.json();
-   return result;
-};
+export {
+   deleteUserPermission,
+   getUsersInWorkspace,
+   postPermission,
+   removeAllUserPermissions
+} from "@/modules/permissions/infrastructure";
+export {
+   confirmRegistration,
+   deleteUser,
+   downloadAvatar,
+   downloadAvatarCid,
+   downloadUserSettings,
+   forgotPassword,
+   loginUser,
+   logout,
+   registerUser,
+   resetPassword,
+   updateUserName,
+   updateUserSettings
+} from "@/modules/users/infrastructure";
 
 export async function getHealth() {
    const url = `${HEALTH_ENDPOINT}`;
@@ -340,61 +320,6 @@ export async function getHealth() {
    const data = await response.json();
    return data;
 }
-
-export const postPermission = async (formData: {
-   email: string;
-   workspaceId: string;
-}) => {
-   const options: RequestInit = {
-      method: "POST",
-      headers: {
-         "Content-Type": "application/json"
-      },
-      body: JSON.stringify({
-         email: formData.email,
-         workspaceId: formData.workspaceId
-      }),
-      credentials: "include"
-   };
-   const response = await fetchWithCredentials(PERMISSIONS_ENDPOINT, withCsrf(options));
-   if (!response.ok) {
-      throw new Error("Failed to add user to the workspace");
-   }
-   const data = await response.json();
-   return data;
-};
-
-export const getUsersInWorkspace = async (workspaceId: string | string[]) => {
-   const url = `${PERMISSIONS_ENDPOINT}/users-in-workspace/${workspaceId}`;
-   const options: RequestInit = {
-      method: "GET",
-      headers: {
-         "Content-Type": "application/json"
-      },
-      credentials: "include"
-   };
-   const response = await fetchWithCredentials(url, options);
-   if (!response.ok) {
-      throw new Error("Failed to add user to the workspace");
-   }
-   const data = await response.json();
-   return data;
-};
-
-export const deleteUserPermission = async (permissionId: number) => {
-   const url = `${PERMISSIONS_ENDPOINT}/users-in-workspace/remove/${permissionId}`;
-   const options: RequestInit = {
-      method: "DELETE",
-      headers: {
-         "Content-Type": "application/json"
-      },
-      credentials: "include"
-   };
-   const response = await fetchWithCredentials(url, withCsrf(options));
-   if (!response.ok) {
-      throw new Error("Failed to remove user from the workspace");
-   }
-};
 
 export function useDocumentsStatistics() {
    const { data, error, isLoading, isValidating, mutate } = useSWR(
@@ -465,255 +390,3 @@ export function useRecentChats() {
       mutate
    };
 }
-
-export const logout = async (): Promise<{
-   status: string;
-   message: string;
-}> => {
-   try {
-      const response = await fetchWithCredentials(
-         `${USERS_ENDPOINT}/logout`,
-         withCsrf({
-            method: "POST",
-            credentials: "include"
-         })
-      );
-
-      if (!response.ok) {
-         throw new Error("Failed to log out");
-      }
-
-      return await response.json();
-   } catch (error) {
-      console.error("Error during logout:", error);
-      throw error;
-   }
-};
-
-export const confirmRegistration = async (
-   token: string,
-   formData: {
-      lang: string;
-      confirmationLink: string;
-   }
-): Promise<{
-   status: string;
-   message: string;
-}> => {
-   try {
-      const url = `${USERS_ENDPOINT}/confirm-registration?token=${token}`;
-      const options: RequestInit = {
-         method: "POST",
-         headers: {
-            "Content-Type": "application/json"
-         },
-         body: JSON.stringify({
-            lang: formData.lang,
-            confirmationLink: formData.confirmationLink
-         }),
-         credentials: "include"
-      };
-      const response = await fetchWithCredentials(url, withCsrf(options));
-
-      return await response.json();
-   } catch (error) {
-      console.error("Error during registration confirmation:", error);
-      throw error;
-   }
-};
-
-export const updateUserSettings = async (formData: FormData) => {
-   try {
-      const res = await fetchWithCredentials(
-         `${USERS_ENDPOINT}/user-settings`,
-         withCsrf({
-            method: "POST",
-            credentials: "include",
-            body: formData
-         })
-      );
-
-      if (!res.ok) {
-         throw new Error("Failed to update user settings");
-      }
-      return res.json();
-   } catch (error) {
-      console.error("Error updating user settings:", error);
-      throw error;
-   }
-};
-
-export const updateUserName = async (name: string) => {
-    try {
-        const res = await fetchWithCredentials(
-            `${USERS_ENDPOINT}/reset-name`,
-            withCsrf({
-                method: "POST",
-                credentials: "include",
-                headers: {
-                    "Content-Type": "application/json"
-                },
-                body: JSON.stringify({ name })
-            })
-        );
-
-        if (!res.ok) {
-            throw new Error("Failed to update user name");
-        }
-        return res.json();
-    }
-    catch (error) {
-        console.error("Error updating user name:", error);
-        throw error;
-    }
-};
-
-export const deleteUser = async () => {
-    try {
-        const res = await fetchWithCredentials(
-            `${USERS_ENDPOINT}/delete-user`,
-            withCsrf({
-                method: "DELETE",
-                credentials: "include"
-            })
-        );
-
-        if (!res.ok) {
-            throw new Error("Failed to delete user");
-        }
-        const text = await res.text();
-        return text ? JSON.parse(text) : null;
-    }
-    catch (error) {
-        console.error("Error deleting user:", error);
-        throw error;
-    }
-};
-
-export const removeAllUserPermissions = async (email: string) => {
-    try {
-        const res = await fetchWithCredentials(
-            `${PERMISSIONS_ENDPOINT}/user/remove-all/${encodeURIComponent(email)}`,
-            withCsrf({
-                method: "DELETE",
-                credentials: "include"
-            })
-        );
-
-        if (!res.ok) {
-            throw new Error("Failed to remove user permissions");
-        }
-        const text = await res.text();
-        return text ? JSON.parse(text) : null;
-    }
-    catch (error) {
-        console.error("Error removing user permissions:", error);
-        throw error;
-    }
-};
-
-export const downloadAvatarCid = async () => {
-   try {
-      const res = await fetchWithCredentials(`${USERS_ENDPOINT}/avatar-cid`, {
-         method: "GET",
-         credentials: "include"
-      });
-
-      if (res.status === 404) {
-         return null;
-      }
-
-      if (!res.ok) {
-         throw new Error("Failed to download avatar CID");
-      }
-
-      const result = await res.json();
-      return result.cid;
-   } catch (error) {
-      console.error("Error downloading avatar CID:", error);
-      throw error;
-   }
-}
-
-export const downloadAvatar = async (avatarCid: string) => {
-   try {
-      const res = await fetchWithCredentials(`${USERS_ENDPOINT}/avatar/${encodeURIComponent(avatarCid)}`, {
-         method: "GET",
-         credentials: "include"
-      });
-
-      if (res.status === 404) {
-         // Avatar not found is expected for new users — return null
-         return null;
-      }
-
-      if (!res.ok) {
-         throw new Error("Failed to download avatar");
-      }
-
-      return res;
-   } catch (error) {
-      console.error("Error downloading avatar:", error);
-      throw error;
-   }
-};
-
-export const downloadUserSettings = async () => {
-   try {
-      const res = await fetchWithCredentials(`${USERS_ENDPOINT}/user-settings`, {
-         method: "GET",
-         credentials: "include"
-      });
-
-      if (res.status === 404) {
-         return {};
-      }
-
-      if (!res.ok) {
-         throw new Error("Failed to download user settings");
-      }
-
-      const result = await res.json();
-      return result;
-   } catch (error) {
-      console.error("Error downloading user settings:", error);
-      throw error;
-   }
-};
-
-export const forgotPassword = async (data: Record<string, string>) => {
-   const url = `${USERS_ENDPOINT}/forgot-password`;
-   const options: RequestInit = {
-      method: "POST",
-      headers: {
-         "Content-Type": "application/json"
-      },
-      body: JSON.stringify({
-         email: data.email,
-         resetPasswordLink: data.resetPasswordLink,
-         lang: data.lang
-      }),
-      credentials: "include"
-   };
-   const response = await fetchWithCredentials(url, withCsrf(options));
-   const result = await response.json();
-   return result;
-};
-
-export const resetPassword = async (data: Record<string, string>) => {
-   const url = `${USERS_ENDPOINT}/reset-password`;
-   const options: RequestInit = {
-      method: "POST",
-      headers: {
-         "Content-Type": "application/json"
-      },
-      body: JSON.stringify({
-         password: data.password,
-         token: data.token
-      }),
-      credentials: "include"
-   };
-   const response = await fetchWithCredentials(url, withCsrf(options));
-   const result = await response.json();
-   return result;
-};
