+++ frontend/src/modules/permissions/index.ts
@@ -1,1 +1,1 @@
-export {};
+export * from "./infrastructure";
