+++ frontend/src/modules/permissions/infrastructure/index.ts
@@ -1,1 +1,1 @@
-export {};
+export * from "./permissions.api";
