+++ frontend/src/modules/users/index.ts
@@ -1,1 +1,1 @@
-export {};
+export * from "./infrastructure";
