+++ frontend/src/modules/users/infrastructure/index.ts
@@ -1,1 +1,1 @@
-export {};
+export * from "./users.api";
